+++ parsley-core/src/ghc-8.10+/Parsley/Internal/Backend/Machine/PosOps.hs
@@ -1,25 +1,34 @@
-{-# LANGUAGE CPP, MagicHash, UnboxedTuples #-}
+{-# LANGUAGE CPP, MagicHash, UnboxedTuples, NumericUnderscores #-}
 module Parsley.Internal.Backend.Machine.PosOps (module Parsley.Internal.Backend.Machine.PosOps) where
-
-import Parsley.Internal.Backend.Machine.Types.Base (Pos)
-import Parsley.Internal.Common                     (Code)
-import GHC.Prim                                    (plusWord#, minusWord#, and#, or#)
-
 
 #include "MachDeps.h"
 #if WORD_SIZE_IN_BITS < 64
 #define FULL_WIDTH_POSITIONS
 #endif
 
+import Parsley.Internal.Backend.Machine.Types.Base (Pos)
+import Parsley.Internal.Common                     (Code)
+import GHC.Exts                                    (Int(..))
+import GHC.Prim                                    (plusWord#, and#, or#, word2Int#, uncheckedShiftRL#)
+#ifdef FULL_WIDTH_POSITIONS
+import GHC.Prim                                    (minusWord#)
+#endif
+
 initPos :: Code Pos
+{-# INLINEABLE updatePos# #-}
 updatePos# :: Pos -> Char -> Pos
+extractLine :: Code Pos -> Code Int
+extractCol :: Code Pos -> Code Int
 
 #ifndef FULL_WIDTH_POSITIONS
-initPos = [|| 0x00010001## ||]
+initPos = [|| 0x00000001_00000001## ||]
 
-updatePos# pos '\n' = (pos `plusWord#` 0x10000##) `and#` 0xffff0001##
-updatePos# pos '\t' = ((pos `plusWord#` 3##) `and#` (0## `minusWord#` 4##)) `or#` 1##
-updatePos# pos _    = pos `plusWord#` 1##
+updatePos# pos '\n' = (pos `plusWord#` 0x00000001_00000000##) `and#` 0xffffffff_00000001##
+updatePos# pos '\t' = ((pos `plusWord#` 0x00000000_00000003##) `and#` 0xffffffff_fffffffc##) `or#` 0x00000000_00000001##
+updatePos# pos _    = pos `plusWord#` 0x00000000_00000001##
+
+extractLine qpos = [||I# (word2Int# ($$qpos `uncheckedShiftRL#` 32#))||]
+extractCol qpos = [||I# (word2Int# ($$qpos `and#` 0x00000000_ffffffff##))||]
 
 #else
 initPos = [|| (# 1##, 1## #) ||]
@@ -27,6 +36,9 @@ initPos = [|| (# 1##, 1## #) ||]
 updatePos# (# line, _ #)   '\n' = (# line `plusWord#` 1##, 1## #)
 updatePos# (# line, col #) '\t' = (# line, ((col `plusWord#` 3##) `and#` (0## `minusWord#` 4##)) `or#` 1## #) -- nearest tab boundary `c + (4 - (c - 1) % 4)`
 updatePos# (# line, col #) _    = (# line, col `plusWord#` 1## #)
+
+extractLine qpos = [|| case $$qpos of (# line, _ #) -> I# (word2Int# line) ||]
+extractCol qpos = [|| case $$qpos of (# _, col #) -> I# (word2Int# col) ||]
 #endif
 
 updatePos :: Code Pos -> Code Char -> Code Pos